🚚 Rename `src/redux/` folder to `src/store/`
As with every feature folder, the root store folder should be named `store`, to align with Redux best practices.
