+++ src/main.jsx
@@ -2,7 +2,7 @@ import { BrowserRouter } from 'react-router-dom'
 import ReactDOM from 'react-dom/client'
 import { Provider } from 'react-redux'
 import { CssBaseline, ThemeProvider } from '@mui/material'
-import { store } from './redux'
+import { store } from './store'
 import { App } from './components/App/App'
 import { muiONetworkTheme } from './styles/muiONetworkTheme'
 import '@fontsource/roboto/400.css'
